backend/requirements.txt
# FastAPI Backend for EgyptAir Feedback System
# Install with: pip install -r requirements.txt

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1

# NLP & Sentiment Analysis
nltk==3.8.1
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99

# Arabic NLP (optional - comment out if installation fails)
# camel-tools==1.5.7

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# CORS
starlette==0.35.1

# Testing
pytest==7.4.4
httpx==0.26.0
